created search bar component
