+++ lib/dpul_collections_web/components/layouts/app.html.heex
@@ -1,3 +1,4 @@
+<.live_component module={DpulCollectionsWeb.SearchBarComponent} id="search-bar" />
 <main id="main-content" class="max-w-7xl mx-auto py-6 px-4 sm:px-6 lg:px-8">
   <.flash_group flash={@flash} />
   {@inner_content}
