+++ lib/dpul_collections_web/live/home_live.ex
@@ -20,16 +20,6 @@ defmodule DpulCollectionsWeb.HomeLive do
   def render(assigns) do
     ~H"""
     <div class="grid grid-flow-row auto-rows-max gap-20">
-      <div>
-        <form phx-submit="search">
-          <div class="grid grid-cols-4">
-            <input class="col-span-4 md:col-span-3" type="text" name="q" value={@q} />
-            <button class="col-span-4 md:col-span-1 btn-primary" type="submit">
-              {gettext("Search")}
-            </button>
-          </div>
-        </form>
-      </div>
       <div id="welcome" class="grid place-self-center gap-10 max-w-prose">
         <h3 class="text-5xl text-center">{gettext("Explore Our Digital Collections")}</h3>
         <p class="text-xl text-center">
@@ -51,11 +41,5 @@ defmodule DpulCollectionsWeb.HomeLive do
       </div>
     </div>
     """
-  end
-
-  def handle_event("search", %{"q" => q}, socket) do
-    params = %{q: q} |> Helpers.clean_params()
-    socket = push_navigate(socket, to: ~p"/search?#{params}")
-    {:noreply, socket}
   end
 end
