+++ test/dpul_collections_web/features/search_bar_test.exs
@@ -11,11 +11,14 @@ defmodule DpulCollectionsWeb.Features.SearchBarTest do
     on_exit(fn -> Solr.delete_all(active_collection()) end)
   end
 
-  test "submitting a search on the search page", %{conn: conn} do
+  test "submitting a search via the search bar component on the search page", %{conn: conn} do
     conn
     |> visit("/search")
     |> fill_in("Search", with: "Document-3")
-    |> click_button("Search")
+    |> within(".search-bar", fn session ->
+      session
+      |> click_button("Search")
+    end)
     |> assert_has("#item-counter", text: "1 - 1 of 1")
   end
